+++ resources/views/layouts/app.blade.php
@@ -54,5 +54,7 @@
                 </x-slot>
             </x-dialog>
         </div>
+
+        <x-autosave-draft-recovery-modal />
     </body>
 </html>